infrastructure/support_orchestrator/source/pyproject.toml
[tool.poetry]
name = "support_orchestrator"
version = "0.1.0"
description = ""
authors = ["Piotr Tutak <p.tutak@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Flask = "^1.1.2"
py4j = "^0.10.9"
waitress = "^2.0.0"
PyYAML = "^5.4.1"
phoenixdb = "^1.0.1"
requests = "^2.25.1"
hdfs = "^2.6.0"
Cerberus = "^1.3.4"
psutil = "^5.8.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.4"
flake8-black = "^0.2.1"
isort = "^5.7.0"
flake8-isort = "^4.0.0"
black = "^20.8b1"
pytest-flask = "^1.2.0"
importlib-metadata = "1.7.0"
mypy = "^0.910"

[tool.flake8]
max-line-length = 120

[tool.isort]
profile = "black"

[tool.black]
line-length = 120


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
